+++ anubis-core/src/main/java/com/anubis/core/service/auth/AuthService.java
@@ -1,0 +1,9 @@
+package com.anubis.core.service.auth;
+
+
+import com.anubis.core.dto.LoginRequest;
+import com.anubis.core.dto.UserInfoResponse;
+
+public interface AuthService {
+    UserInfoResponse authenticateUser(LoginRequest request);
+}
