+++ anubis-core/src/main/java/com/anubis/core/service/auth/AuthServiceImpl.java
@@ -1,16 +1,17 @@
-package com.anubis.family.api.service.auth;
+package com.anubis.core.service.auth;
 
+import com.anubis.core.dto.LoginRequest;
+import com.anubis.core.dto.UserInfoResponse;
 import com.anubis.core.exception.UsernamePasswordAuthenticationException;
-import com.anubis.family.api.model.request.LoginRequest;
-import com.anubis.family.api.model.response.UserInfoResponse;
-import com.anubis.family.api.service.user.UserDetailsImpl;
-import com.anubis.family.api.util.JwtUtil;
+import com.anubis.core.util.JwtHelper;
+import com.anubis.core.service.user.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseCookie;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -25,7 +26,7 @@ public class AuthServiceImpl implements AuthService {
     AuthenticationManager authenticationManager;
 
     @Autowired
-    JwtUtil jwtUtil;
+    JwtHelper jwtHelper;
 
     @Override
     public UserInfoResponse authenticateUser(LoginRequest loginRequest)
@@ -46,7 +47,7 @@ public class AuthServiceImpl implements AuthService {
         ResponseCookie jwtCookie = getJwtUtil().generateJwtCookie(userDetails);
         List<String> roles = userDetails.getAuthorities()
                 .stream()
-                .map(item -> item.getAuthority())
+                .map(GrantedAuthority::getAuthority)
                 .collect(Collectors.toList());
 
         UserInfoResponse response = new UserInfoResponse();
@@ -68,11 +69,11 @@ public class AuthServiceImpl implements AuthService {
         this.authenticationManager = authenticationManager;
     }
 
-    public JwtUtil getJwtUtil() {
-        return jwtUtil;
+    public JwtHelper getJwtUtil() {
+        return jwtHelper;
     }
 
-    public void setJwtUtil(JwtUtil jwtUtil) {
-        this.jwtUtil = jwtUtil;
+    public void setJwtUtil(JwtHelper jwtHelper) {
+        this.jwtHelper = jwtHelper;
     }
 }
