+++ anubis-core/src/main/java/com/anubis/core/service/user/UserDetailsImpl.java
@@ -1,6 +1,6 @@
-package com.anubis.family.api.service.user;
+package com.anubis.core.service.user;
 
-import com.anubis.family.api.model.User;
+import com.anubis.core.entity.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
