+++ anubis-family-api/src/main/java/com/anubis/family/api/controller/DailyTaskController.java
@@ -2,9 +2,8 @@ package com.anubis.family.api.controller;
 
 import com.anubis.core.entity.family.DailyTask;
 import com.anubis.core.entity.family.FamilyMember;
-import com.anubis.core.service.task.DailyTaskServiceImpl;
 import com.anubis.core.service.task.TaskServiceImpl;
-import com.anubis.family.api.service.family.FamilyMemberService;
+import com.anubis.core.service.family.FamilyMemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
