+++ anubis-family-api/src/main/java/com/anubis/family/api/controller/TaskController.java
@@ -5,8 +5,7 @@ import com.anubis.core.entity.family.Task;
 import com.anubis.core.service.task.TaskService;
 import com.anubis.family.api.model.IsCompleteDTO;
 import com.anubis.family.api.model.response.MessageResponse;
-import com.anubis.family.api.service.family.FamilyMemberService;
-import com.anubis.family.api.util.JwtUtil;
+import com.anubis.core.service.family.FamilyMemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
