--- anubis-family-api/src/main/java/com/anubis/family/api/model/ProfileInfo.java
@@ -1,11 +1,0 @@
-package com.anubis.family.api.model;
-
-public class ProfileInfo {
-
-    String firstName;
-    String lastName;
-    String email;
-    long userId;
-    long familyMemberId;
-    String username;
-}
