--- anubis-family-api/src/main/java/com/anubis/family/api/service/auth/AuthService.java
@@ -1,8 +1,0 @@
-package com.anubis.family.api.service.auth;
-
-import com.anubis.family.api.model.request.LoginRequest;
-import com.anubis.family.api.model.response.UserInfoResponse;
-
-public interface AuthService {
-    UserInfoResponse authenticateUser(LoginRequest request);
-}
